+++ Development/DeviceTestApp.swift
@@ -12,6 +12,31 @@ struct DeviceTestApp: App {
         Device.current.disableIdleTimerWhenPluggedIn()
         Application.track()
         //Application.appleID = "6736626499"
+        
+        // use this to figure out which items are causing hash conflicts.
+//        var dictionary: [Device: String] = [:]
+//        for device in Device.all {
+//            debug(device.officialName)
+//            if dictionary.keys.contains(device) {
+//                debug("DUPLICATE KEY ALREADY EXISTS!", level: .ERROR)
+//                debug(device)
+//                debug(dictionary.keys.first { $0 == device } ?? "NONE existing using == matching")
+//            }
+//            dictionary[device] = device.officialName
+//        }
+//        var ids = [String: Device]()
+//        for device in Device.all {
+//            if let image = device.image {
+//                debug(image)
+//                if ids.keys.contains(image) {
+//                    debug("DUPLICATE KEY ALREADY EXISTS!", level: .ERROR)
+//                    debug(device)
+//                    debug(ids.keys.first { $0 == image } ?? "NONE existing using == matching")
+//                }
+//                ids[image] = device
+//            }
+//        }
+//        debug("Done with dictionary \(dictionary.count)")
     }
     var body: some Scene {
         WindowGroup {