+++ Package.swift
@@ -7,15 +7,15 @@
 
 import PackageDescription
 
-let version = "2.6.5"
+let version = "2.7.0"
 let packageLibraryName = "Device"
 
 // Products define the executables and libraries a package produces, making them visible to other packages.
 var products = [
 	Product.library(
-        name: "\(packageLibraryName) Library", // has to be named different from the iOSApplication or Swift Playgrounds won't open correctly
-        targets: [packageLibraryName]
-    ),
+		name: "\(packageLibraryName) Library", // has to be named different from the iOSApplication or Swift Playgrounds won't open correctly
+		targets: [packageLibraryName]
+	),
 ]
 
 // Targets are the basic building blocks of a package, defining a module or a test suite.
@@ -23,14 +23,14 @@ var products = [
 var targets = [
 	Target.target(
 		name: packageLibraryName,
-        dependencies: [
-            .product(name: "Color Library", package: "color"), // apparently needs to be lowercase.  Also note this is "Color Library" not "Color"
-        ],
+		dependencies: [
+			.product(name: "Color Library", package: "color"), // apparently needs to be lowercase.  Also note this is "Color Library" not "Color"
+		],
 		path: "Sources"
 		// If resources need to be included in the module, include here
 		,resources: [ // unfortuantely cannot be conditionally compiled based on Swift version since the tool seems to be run on latest version.
-            Resource.process("Resources"),
-        ]
+			Resource.process("Resources"),
+		]
 //		,swiftSettings: [
 //			.enableUpcomingFeature("BareSlashRegexLiterals")
 //		]
@@ -45,27 +45,32 @@ var platforms: [SupportedPlatform] = [
 
 #if canImport(PlaygroundSupport)
 platforms += [
-    .iOS("15.2"), // minimum for Swift Playgrounds support
+	.iOS("15.2"), // minimum for Swift Playgrounds support (maximum version for test iPhone 7)
 ]
 #else
 platforms += [
-    .iOS("11"), // 13 minimum for Combine/SwiftUI, 15 minimum for Date.now
+	.iOS("11"), // 13 minimum for Combine/SwiftUI, 15 minimum for Date.now, (maximum version for test iPhone 7)
 ]
 #endif
 
+#if compiler(>=5.9)
 #if os(visionOS)
 platforms += [
-    .visionOS("1.0"), // unavailable in Swift Playgrounds so has to be separate
+	.visionOS("1.0"), // unavailable in Swift Playgrounds so has to be separate
 ]
+#endif
 #endif
 
 #if canImport(AppleProductTypes) // swift package dump-package fails because of this
 import AppleProductTypes
 
+let executableTargetName = "\(packageLibraryName)TestAppModule"
+
 products += [
 	.iOSApplication(
-		name: packageLibraryName, // needs to match package name to open properly in Swift Playgrounds
-		targets: ["\(packageLibraryName)TestAppModule"],
+		name: "\(packageLibraryName) App", // needs to match package name to open properly in Swift Playgrounds <v4.5, but must be different to run in v4.6 and greater.
+		targets: [executableTargetName],
+//		bundleIdentifier: "com.kudit.compatibility", // ignored in playgrounds
 		teamIdentifier: "3QPV894C33",
 		displayVersion: version,
 		bundleVersion: "1",
@@ -81,16 +86,16 @@ products += [
 			.landscapeLeft,
 			.portraitUpsideDown(.when(deviceFamilies: [.pad]))
 		],
-        capabilities: [
-            .outgoingNetworkConnections() // for networking tests and loading device images
-        ],
+		capabilities: [
+			.outgoingNetworkConnections() // for networking tests and loading device images
+		],
 		appCategory: .developerTools
 	),
 ]
 
 targets += [
 	.executableTarget(
-		name: "\(packageLibraryName)TestAppModule",
+		name: executableTargetName,
 		dependencies: [
 			.init(stringLiteral: packageLibraryName), // have to use init since normally would be assignable by string literal but we're not using a string literal
 		],
@@ -116,13 +121,13 @@ targets += [
 #endif // for Swift Package compiling for https://swiftpackageindex.com/add-a-package
 
 let package = Package(
-    name: packageLibraryName,
-    platforms: platforms,
-    products: products,
-    // include dependencies
-    dependencies: [
-        // Dependencies declare other packages that this package depends on.
-        .package(url: "https://github.com/kudit/Color", "1.1.4"..<"2.0.0"),
-    ],
-    targets: targets
+	name: packageLibraryName,
+	platforms: platforms,
+	products: products,
+	// include dependencies
+	dependencies: [
+		// Dependencies declare other packages that this package depends on.
+		.package(url: "https://github.com/kudit/Color", "1.1.4"..<"2.0.0"),
+	],
+	targets: targets
 )