+++ Sources/Screen.swift
@@ -91,11 +91,11 @@ public struct Screen: Hashable, Sendable {
     public static let i61x1125 = Screen(diagonal: 6.1, resolution: (1125,2436), ppi: 458)
     // iPhone 12 Pro, 13 Pro, 14, 16e
     public static let i61x1170 = Screen(diagonal: 6.1, resolution: (1170,2532), ppi: 460)
-    // iPhone 14 Pro
+    // iPhone 14 Pro, 15, 15 Pro, 16
     public static let i61x1179 = Screen(diagonal: 6.1, resolution: (1179,2556), ppi: 460)
-    // iPhone 12 Pro Max, iPhone 14 Plus
+    // iPhone 12 Pro Max, 14 Plus
     public static let i67x1284 = Screen(diagonal: 6.7, resolution: (1284,2778), ppi: 458)
-    // iPhone 14,15 Pro Max
+    // iPhone 15 Pro Max, 16 Plus
     public static let i67x1290 = Screen(diagonal: 6.7, resolution: (1290,2796), ppi: 460)
     // iPhone 16 Pro
     public static let i63 = Screen(diagonal: 6.3, resolution: (1206,2622), ppi: 460)